+++ CHANGELOG.md
@@ -124,6 +124,7 @@ However, if the shopkeeper actually moved from its previous spawn location and n
   * We no longer add the nameplate prefix if the NPC is of type player, because this messes with the NPC's skin.
   * When setting the NPC's name we now also set the nameplate visibility: If the name is empty, the nameplate is hidden. Otherwise, the 'always-show-nameplates' setting determines whether to always show the nameplate, or to only show it when looking directly at the entity (this only works for non-player NPCs though).
 * Config: Added shulker to the by default enabled mob types. However, note that this mob will stay in its closed form currently due to its disabled AI.
+* Added an editor option to change the color of shulker shopkeepers.
 
 API:  
 * PlayerCreatePlayerShopkeeperEvent and PlayerShopkeeperHireEvent: The meaning of the max shops limit has changed. A value of 0 or less no longer indicates 'no limit'.
@@ -228,6 +229,8 @@ Messages:
 * Added 'villager-name-invalid'.
 * Added 'button-snowman-pumpkin-head'.
 * Added 'button-snowman-pumpkin-head-lore'.
+* Added 'button-shulker-color'.
+* Added 'button-shulker-color-lore'.
 * Minor changes to the german translation.
 
 You will have to manually update your custom language files to adapt for these changes.